+++ src/ralph/api/auth/user.py
@@ -1,6 +1,6 @@
 """Authenticated user for the Ralph API."""
 
-from typing import Dict, FrozenSet, Literal
+from typing import Dict, FrozenSet, Literal, Optional
 
 from pydantic import BaseModel
 
@@ -67,7 +67,9 @@ class AuthenticatedUser(BaseModel):
     Attributes:
         agent (dict): The agent representing the current user.
         scopes (list): The scopes the user has access to.
+        target (str or None): The target index or database to store statements into.
     """
 
     agent: Dict
     scopes: UserScopes
+    target: Optional[str]