+++ tests/backends/lrs/test_es.py
@@ -273,7 +273,7 @@ def test_backends_lrs_es_query_statements_query(
     parameters, should produce the expected Elasticsearch query.
     """
 
-    def mock_read(query, chunk_size):
+    def mock_read(query, target, chunk_size):
         """Mock the `ESLRSBackend.read` method."""
         assert query.dict() == expected_query
         assert chunk_size == expected_query.get("size")
@@ -295,18 +295,32 @@ def test_backends_lrs_es_query_statements(es, es_lrs_backend):
     """Test the `ESLRSBackend.query_statements` method, given a query,
     should return matching statements.
     """
+    # Create a custom index
+    custom_target = "custom-target"
+    es.indices.create(index=custom_target)
 
     # Instantiate ESLRSBackend.
     backend = es_lrs_backend()
-    # Insert documents.
-    documents = [{"id": "2", "timestamp": "2023-06-24T00:00:20.194929+00:00"}]
-    assert backend.write(documents) == 1
+    # Insert documents into default target.
+    documents_default = [{"id": "2", "timestamp": "2023-06-24T00:00:20.194929+00:00"}]
+    assert backend.write(documents_default) == 1
+    # Insert documents into custom target.
+    documents_custom = [{"id": "3", "timestamp": "2023-05-25T00:00:20.194929+00:00"}]
+    assert backend.write(documents_custom, target=custom_target) == 1
 
     # Check the expected search query results.
     result = backend.query_statements(RalphStatementsQuery.construct(limit=10))
-    assert result.statements == documents
+    assert result.statements == documents_default
     assert re.match(r"[0-9]+\|0", result.search_after)
 
+    # Check the expected search query results on custom target.
+    result = backend.query_statements(
+        RalphStatementsQuery.construct(limit=10), target=custom_target
+    )
+    assert result.statements == documents_custom
+    assert re.match(r"[0-9]+\|0", result.search_after)
+
+    es.indices.delete(index=custom_target)
     backend.close()
 
 
@@ -400,6 +414,14 @@ def test_backends_lrs_es_query_statements_by_ids_with_multiple_indexes(
     assert not list(backend_1.query_statements_by_ids(["2"]))
     assert not list(backend_2.query_statements_by_ids(["1"]))
     assert list(backend_2.query_statements_by_ids(["2"])) == [index_2_document]
+
+    # Check that backends can also read from another target
+    assert list(
+        backend_1.query_statements_by_ids(["2"], target=ES_TEST_FORWARDING_INDEX)
+    ) == [index_2_document]
+    assert list(backend_2.query_statements_by_ids(["1"], target=ES_TEST_INDEX)) == [
+        index_1_document
+    ]
 
     backend_1.close()
     backend_2.close()